# apps/dashboard/server/src/openapi.yaml
openapi: 3.0.3
info:
  title: WarrantDB API
  version: 1.0.0
...
servers:
  - url: http://localhost:8080/api
paths:
...
  /auth/login:
    post:
      summary: Sign in with email and password, establishing a server session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Session cookie set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '401':
          description: Invalid email or password
...
  /auth/forgot-password:
    post:
      summary: Request a password-reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '202':
          description: Accepted regardless of whether the email has an account (avoids leaking account existence)
  /auth/reset-password:
    post:
      summary: Set a new password using a reset/invite token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Password updated
        '400':
          description: Token is invalid, expired, or password too short
  /auth/me:
    get:
      summary: Current authenticated user profile
      responses:
        '200':
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthProfileResponse'
        '401':
          description: Not authenticated
  /auth/logout:
    post:
      summary: Clear session cookie and sign out
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '401':
          description: Not authenticated
...
  /auth/access-request:
    post:
      summary: Submit an access request
      description: Allows prospective users to notify administrators that they need an account. No authentication required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                displayName:
                  type: string
                message:
                  type: string
      responses:
        '202':
          description: Request accepted for review
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  request:
                    $ref: '#/components/schemas/AccessRequest'
        '400':
          description: Validation error
        '409':
          description: Account already exists
...
components:
  schemas:
...
    AuthUser:
      type: object
      properties:
        uid: { type: string }
        email: { type: string, nullable: true }
        emailVerified: { type: boolean }
        displayName: { type: string, nullable: true }
        roles:
          type: array
          items: { type: string }
        departments:
          type: array
          items: { type: string }
        counties:
          type: array
          items: { type: string }
        status:
          type: string
        mfaEnforced:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AuthSessionResponse:
      type: object
      properties:
        ok: { type: boolean }
        user: { $ref: '#/components/schemas/AuthUser' }
        sessionExpiresAt:
          type: string
          format: date-time
    AuthProfileResponse:
      type: object
      properties:
        ok: { type: boolean }
        user: { $ref: '#/components/schemas/AuthUser' }
        offense: { type: string }
        case_number: { type: string }
...
    AccessRequest:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        displayName: { type: string, nullable: true }
        message: { type: string, nullable: true }
        status: { type: string, enum: ['pending','reviewed','completed','rejected'] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...